+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add text based status to security assessment generator
 - Move generation to be a background task to avoid idle timeouts
 - Swapped progress spinner to something more infoblox
+- Move to new authentication classes
 
 ## v0.4.6
 - Fix bug where invalid API Keys would not always be reported